refactor: Improved rendering and state monitoring of selected video
This is a small feature improvement to the component which displays
the selected video. Instead of passing a literal video object to the
VideoSelected component of the app, I added another state property
to capture a selected video from the response data from the AJAX
request. This will be passed to the component which will re-render
once data is available.

TODO: Add video selection to each video list item component.