feat: Added video selection to each list item in video list
Added functionality to VideoListItem component so that each list item
in VideoList will have an on-click event listener that triggers a
function to capture the video object within the VideoListItem
and change the state of the App component's selectedVideo property.
This selectedVideo's state will then be transferred to the
VideoSelected component which renders the video and its details for
viewing.