feat: Change styling and added video search functionality
This commit made a few changes to the overall styling of some
components of the app for better alignment and visual improvements.
Search functionality was also added to the application. The YTSearch
from YouTube API Search library was refactored to run inside a method.
This method is triggered with search term parameters from the SearchBar
component. All search terms were also checked to ensure that only puppy
videos are provided to make this application exclusively show puppy
videos.